multiple_agents/pyproject.toml
[project]
name = "multiple-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.14",
    "geopy>=2.4.1",
    "google-adk>=1.7.0",
    "gradio>=5.38.0",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "mcp>=1.12.0",
    "python-dotenv>=1.1.1",
]
